requirements.txt
pika==1.0.0
django==2.1.5
gensim==3.4.0
jieba==0.39
pyyaml==5.1
numpy==1.16.2
scipy==1.2.1
scikit-learn==0.20.3
torch==1.0.0
requests==2.21.0